move git utilities to their own file
